Update Auth JWT
