+++ .env
@@ -2,9 +2,3 @@ APP_SECRET=15a0dfa98deda3db69e61fca885c1db4
 ###> nelmio/cors-bundle ###
 CORS_ALLOW_ORIGIN=^https?://localhost(:[0-9]+)?$
 ###< nelmio/cors-bundle ###
-
-###> lexik/jwt-authentication-bundle ###
-JWT_SECRET_KEY=%kernel.project_dir%/config/jwt/private.pem
-JWT_PUBLIC_KEY=%kernel.project_dir%/config/jwt/public.pem
-JWT_PASSPHRASE=cd5ba4a0f619c7612642234e0b05fdfc
-###< lexik/jwt-authentication-bundle ###
