+++ config/packages/api_platform.yaml
@@ -18,3 +18,8 @@ api_platform:
         shared_max_age: 3600
         vary: ['Content-Type', 'Authorization']
         public: true
+    swagger:
+        api_keys:
+           apiKey:
+              name: Authorization
+              type: header
