+++ config/routes.yaml
@@ -1,3 +1,2 @@
-authentication_token:
-    path: /authentication_token
-    methods: ['POST']
+api_login_check:
+    path: /login_check
